+++ components/datepicker/index.jsx
@@ -20,7 +20,7 @@ function createPicker(TheCalendar) {
         },  // onChange 可用于 Validator
         locale: {},
         align: {
-          offset: [0, -10],
+          offset: [0, -9],
         },
         open: false
       };
@@ -87,24 +87,25 @@ function createPicker(TheCalendar) {
       this.props.onChange(timeValue);
     },
     render() {
+      const locale = this.getLocale();
       // 以下两行代码
       // 给没有初始值的日期选择框提供本地化信息
       // 否则会以周日开始排
-      let defaultCalendarValue = new GregorianCalendar(this.getLocale());
+      let defaultCalendarValue = new GregorianCalendar(locale);
       defaultCalendarValue.setTime(Date.now());
 
       const placeholder = ('placeholder' in this.props)
-        ? this.placeholder : this.getLocale().lang.placeholder;
+        ? this.props.placeholder : locale.lang.placeholder;
       const calendar = (
         <TheCalendar
           disabledDate={this.props.disabledDate}
-          locale={this.getLocale().lang}
+          locale={locale.lang}
           defaultValue={defaultCalendarValue}
           dateInputPlaceholder={placeholder}
           showTime={this.props.showTime}
           prefixCls="ant-calendar"
           showOk={this.props.showTime}
-          showClear={true}/>
+          showClear />
       );
 
       let sizeClass = '';
@@ -139,6 +140,7 @@ function createPicker(TheCalendar) {
                          onChange={this.handleInputChange}
                          value={value && this.getFormatter().format(value)}
                          placeholder={placeholder}
+                         style={this.props.style}
                          className={'ant-calendar-picker-input ant-input' + sizeClass}/>
                   <span className="ant-calendar-picker-icon"/>
                 </span>